requirements.txt
# Web Server
Flask==2.3.3
Flask-CORS==4.0.0

# API Integration
openai==1.3.0
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Utilities
uuid==1.30
python-dateutil==2.8.2

# Development Tools (optional)
pytest==7.4.0
black==23.7.0